refactor: 공통 Message 모델 분리 및 적용
- callLog, converter, response DTO에서 사용하던 Message 클래스를
  common.model 패키지로 이동하여 재사용성 향상
- CallLogDetailResponse 내부 중첩 클래스 제거하고 공통 모델 참조로 변경
