+++ backend/src/main/java/com/airing/backend/common/converter/JsonConverter.java
@@ -2,7 +2,7 @@ package com.airing.backend.common.converter;
 
 import java.util.List;
 
-import com.airing.backend.callLog.dto.CallLogEventRequest.Message;
+import com.airing.backend.common.model.Message;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
