+++ backend/src/main/java/com/airing/backend/common/model/Message.java
@@ -1,0 +1,15 @@
+package com.airing.backend.common.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class Message {
+    private String from;    // 'ai' or 'user'
+    private String message;
+}
